# swagger/sample-contract-0.0.1-snapshot.yaml
swagger: '2.0'
info:
  version: 0.0.1 - snapshot
  title: Contract Sample
  description: >-
    Descrição e informações da api
  contact:
    name: Douglas Gomes
    email: douglastuiuiu@gmail.com.br
basePath: /api/
host: 127.0.0.1:8080
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /v1/projectname/sample:
    post:
      description: Exemplo operação POST
      produces:
        - application/json
      parameters:
        - name: Sample
          in: body
          description: modelo de dados da requisição
          required: true
          schema:
            $ref: '#/definitions/sampleModel'
      responses:
        '200':
          description: Successful response
        '400':
          description: Bad Request / Business Rule
          schema:
            $ref: '#/definitions/errorModel'            
        '401':
          description: Unauthorized        
          schema:
            $ref: '#/definitions/errorModel'
        '500':
          description: Internal Server Error 
          schema:
            $ref: '#/definitions/errorModel'            
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'            
  
  /v1/projectname/sample/{id}:
    get:
      description: Exemplo operação GET
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID do request (_id da collection do mongo)
          required: true
          type: String
      responses:
        '200': 
          description: Successful response
          schema:
            $ref: '#/definitions/testModel'
        '400':
          description: Bad Request / Business Rule
          schema:
            $ref: '#/definitions/errorModel'            
        '401':
          description: Unauthorized        
          schema:
            $ref: '#/definitions/errorModel'
        '500':
          description: Internal Server Error 
          schema:
            $ref: '#/definitions/errorModel'            
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'            

definitions:
  
  sampleModel:
    type: object
    required:
      - id
      - name
      - person
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      person:
        $ref: '#/definitions/person'

  person:
    type: object
    required:
      - id
    properties:
      id:
        type: integer
        format: int64
        description: 'id do usuário'
        
  testModel:
    type: object
    properties:
      id:
        type: integer
        format: int64
      field:
        type: string

  errorModel:
    description: erro genérico padrão
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string